Move AoC and Hacktoberfest into events folder Moves Advent of Code and Hacktoberfest into an events folder. Although these are roughly associated with holidays, they are standalone events that we have participated in in the past.
Therefore they're being moved to an events folder
separate from the "fun" or "holidays" folders.
