+++ bot/exts/events/advent_of_code/_caches.py
