+++ bot/exts/events/hacktoberfest/timeleft.py
