+++ bot/resources/events/advent_of_code/about.json
